decimal to int

+++ Members/Pages/Transparency.cshtml.cs
@@ -33,7 +33,7 @@ namespace Members.Pages
         public int MemberCount => PersonCount + CompanyCount;
 
         public int Employees { get; set; }
-        public int Revenue { get; set; }
+        public decimal Revenue { get; set; }
         public int Associates { get; set; }
 
         public TransparencyModel(ISession session, MemberService mService)
